Add support for vec_rbind(), loose support for subsetting descriptions
Before this commit, vec_rbind() of data frames with resources would loose
resource descriptions (unlike c() or vec_c() for some reason). The culprit
appears to be vec_restore.ldf_resource(). This patch amends that function
such that it no longer filters the description to only those URIs that are
included in the resource. This means that subsetting a resource doesn't
also subset the description. The code has been commented out in the hope
this may soon be reinstated.